% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{data}
\name{acc_dat}
\alias{acc_dat}
\title{The accepted plant names from the Plant List}
\format{
A data frame with 594102 observations on the following 5 variables.
\describe{ 
\item{ID}{a character vector representing the ID of the
name from the plant list database} 
\item{FAMILY}{a character vector
representing the family of this scientific name}
\item{SCIENTIFIC_NAME}{a character vector representing the
scientific name} 
\item{AUTHOR}{a character vector representing the
author of the scientific name} 
\item{STATUS}{a character vector
representing the status of the scientific name} }
}
\source{
The embed database is available at http://pan.baidu.com/s/1hqHrW9I
}
\description{
The data frame contains the columns of the status of a scientific name from
the Plant List.
}
\details{
Please refer to the format section.
}
\examples{

data(acc_dat)

}
\references{
The Plantlist Website \url{http://www.theplantlist.org/}
}
\keyword{datasets}
