\name{cnplants}
\alias{cnplants}
\docType{data}
\title{
Profile data for plant species occurred in China
}
\description{
A curated dataset that contains taxonomic and conservation status for the plant species occurred in China.
}
\usage{data("cnplants")}
\format{
  A data frame with 54617 observations on the following 12 variables.
  \describe{
    \item{\code{SPECIES_CN}}{a character vector}
    \item{\code{SPECIES}}{a character vector}
    \item{\code{SPECIES_FULL}}{a character vector}
    \item{\code{GENUS}}{a character vector}
    \item{\code{GENUS_CN}}{a character vector}
    \item{\code{TPL_FAMILY}}{a character vector}
    \item{\code{FAMILY_ENGLER}}{a character vector}
    \item{\code{TPL_ORDER}}{a character vector}
    \item{\code{IUCN_CHINA}}{a character vector}
    \item{\code{ENDEMIC_TO_CHINA}}{a character vector}
    \item{\code{PROVINTIAL_DISTRIBUTION}}{a character vector}
    \item{\code{ALTITUDE}}{a character vector}
  }
}

\source{
Various sources
}
\references{
To be added
}
\examples{
data(cnplants)
}
\keyword{datasets}
