% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{data}
\name{cnplants_dat}
\alias{cnplants_dat}
\title{The Chinese Plants Database}
\format{
A data frame with 49479 observations on the following 15 variables.
\describe{ 
\item{SPECIES_CN}{a character vector}
\item{SPECIES}{a character vector} 
\item{SPECIES_FULL}{a character vector} 
\item{GENUS}{a character vector}
\item{GENUS_AUTHOR}{a character vector} 
\item{GENUS_CN}{a character vector} 
\item{FAMILY}{a character vector}
\item{FAMILY_CN}{a character vector} 
\item{FAMILY_NUMBER}{a character vector showing the APGIII numbers for Angiosperms, and the
      Christenhusz Family Number for Gymnosperms and ferns} 
\item{ORDER}{a character vector} 
\item{GROUP}{a character vector}
\item{IUCN_CHINA}{a character vector}
\item{ENDEMIC_TO_CHINA}{a character vector}
\item{PROVINTIAL_DISTRIBUTION}{a character vector}
\item{ALTITUDE}{a character vector} }
}
\source{
Compiled from various sources. See
https://github.com/helixcn/plantlist_data
}
\description{
The Chinese Plants Database from Various sources
}
\examples{

data(cnplants_dat)

}
\references{
https://github.com/helixcn/plantlist_data
}
\keyword{datasets}
