% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_taxa.R
\name{count_taxa}
\alias{count_taxa}
\title{Get the number of taxa under each taxonomic level for a dataframe generated
by the function CTPL}
\usage{
count_taxa(checklist_dat)
}
\arguments{
\item{checklist_dat}{A data.frame generated by CTPL or in the same format.}
}
\value{
\item{NO_OF_FAMILIES}{Number of families in the data.frame}

\item{NO_OF_GENERA}{Number of genera in the data.frame}

\item{NO_OF_SPECIES}{Number of species in the data.frame}

\item{NO_OF_FAMILIES_BY_GROUP}{Number of families under each group of higher
plants, namely, Moss, Ferns and Lycophytes, Gymnosperms, Angiosperms}

\item{NO_OF_GENERA_BY_GROUP }{Number of genera under each group, namely,
Moss, Ferns and lycophytes, Gymnosperms, Angiosperms}

\item{NO_OF_SPECIES_BY_GROUP }{Number of species under each group, namely,
Moss, Ferns and lycophytes, Gymnosperms, Angiosperms}

\item{NO_OF_GENERA_BY_FAMILY }{The name is self-explanatory}

\item{NO_OF_SPECIES_BY_FAMILY}{The name is self-explanatory}

\item{NO_OF_SPECIES_BY_GENUS }{The name is self-explanatory}
}
\description{
Compute the number of species under each family, each genus, and each group.
See the value secion.
}
\details{
Simple function summarizing the number of taxa under each taxonomic
level.
}
\examples{

## Do not run
## See the vignette

}
\references{
None
}
\seealso{
\code{\link{make_checklist}}, \code{\link{CTPL}}
}
\author{
Jinlong Zhang
}
