\name{genera_dat}
\alias{genera_dat}
\docType{data}
\title{
The genera-family database for vascular plants
}
\description{
A curated genera-family database for vascular plants mainly based on the Plant List Website \url{http://www.theplantlist.org/1.1/browse/-/-/}
}
\usage{data("genera_dat")}
\format{
  A data frame with 24814 observations on the following 2 variables.
  \describe{
    \item{\code{GENUS}}{a character vector}
    \item{\code{FAMILY}}{a character vector}
  }
}
\source{
the Plant List Website \url{http://www.theplantlist.org/1.1/browse/-/-/}
}
\references{
To add
}
\examples{
data(genera_dat)
}
\keyword{datasets}
