% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_checklist.R
\name{make_checklist}
\alias{make_checklist}
\title{Generate a Markdown Checklist of Species based on the output of CTPL}
\usage{
make_checklist(
  checklist_dat = NULL,
  outfile = "output_make_checklist.md",
  theme = c("complex", "simple", "minimal")
)
}
\arguments{
\item{checklist_dat}{A data.frame generated by CTPL, or the same form.}

\item{outfile}{The name of the markdown file to generate.}

\item{theme}{Different themes, only selected one once.}
}
\value{
This function only save the source code of a checklist of plants to
the working directory. It does not return any value.
}
\description{
Generated a checklist of all the species based on the output of CTPL, with
all the Families appear based on modern classification systems (eg. APGIII).
}
\details{
The output is in Markdown format, which is very convenient to be converted
to tex, word, html, pdf ... using Pandoc or the rmarkdown package.
}
\examples{

# Do not run
# For more information, see the vignette

}
\references{
More details see the vignette.
}
\seealso{
\code{\link{CTPL}}, \code{\link{count_taxa}},
}
\author{
Jinlong Zhang
}
