% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{data}
\name{orders_dat}
\alias{orders_dat}
\title{Order-Family Database for higher plants}
\format{
A data frame with 1879 observations on the following 4 variables.
\describe{ 
\item{FAMILY}{a character vector} 
\item{ORDER}{a character vector} 
\item{FAMILY_NUMBER}{a character vector}
\item{GROUP}{a character vector} }
}
\source{
http://www.mobot.org/MOBOT/research/APweb/top/synonymyfamilies.html

For bryophytes
http://duocet.ibiodiversity.net/index.php?title=%E5%A4%9A%E8%AF%86%E8%8B%94%E8%97%93%E7%B3%BB%E7%BB%9F
}
\description{
Order-Family Database for higher plants, including bryophytes, ferns and
lycophytes, gymnosperms and angiosperms
}
\examples{

data(orders_dat)


}
\references{
Mark W. Chase, Mark W. Chase and James L. Reveal. 2009. A Phylogenetic
Classification of the Land Plants to Accompany APG III. Botanical Journal of
the Linnean Society 161(2):122-127.
}
\keyword{datasets}
