% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{data}
\name{syn_dat}
\alias{syn_dat}
\title{Synonyms database from The Plant List}
\format{
A data frame with 703949 observations on the following 6 variables.
\describe{ 
\item{ID}{a character vector representing the ID in the
Plant List Database} 
\item{FAMILY}{a character vector representing
the Family of the species} 
\item{SCIENTIFIC_NAME}{a character vector
representing the scientific name} 
\item{AUTHOR}{a character vector
representing the authorship of this scientific name} 
\item{STATUS}{a
character vector representing the validity of this scientific name}
\item{ACCEPTED_ID}{a character vector representing the ID of the
accepted name for this scientific name} }
}
\source{
The embbed database is available at http://pan.baidu.com/s/1hqHrW9I
}
\description{
This data frame contains the synonyms for the scientific names from the
Plant List.
}
\details{
Please refer to the usage section
}
\examples{

data(syn_dat)

}
\references{
The Plantlist Website 
\url{http://www.theplantlist.org/}
}
\keyword{datasets}
