% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.table.R
\name{taxa.table}
\alias{taxa.table}
\title{Making a Taxa Table based on the result of function TPL for Phylomatic}
\usage{
taxa.table(x, file = NULL, substitute.sp.white.space = "_")
}
\arguments{
\item{x}{A data.frame generated by function \code{\link{TPL}}}

\item{file}{A character string referring the name of the plain text file to
be generated.}

\item{substitute.sp.white.space}{The substitution symbol between the genus
and specific epithet.}
}
\value{
Taxa table for software Phylomatic.
}
\description{
Making taxa table for software Phylomatic, as "Family/Genus/species"
}
\details{
x must be a data.frame generated by function \code{\link{TPL}}.
}
\examples{


sp <- c( "Ranunculus japonicus", "Anemone udensis", 
         "Ranunculus repens", "Ranunculus chinensis", 
         "Solanum nigrum", "Punica sp." ) 
res <- TPL(sp)
taxa.table(res)


}
\references{
Webb, C. O. and M. J. Donoghue (2005). Phylomatic: tree assembly
for applied phylogenetics. Molecular Ecology Notes 5(1): 181-183.
}
\seealso{
\code{\link{TPL}}
}
\author{
Jinlong Zhang \email{ jinlongzhang01@gmail.com }
}
