% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{data}
\name{genera_dat}
\alias{genera_dat}
\title{The genera-family database for vascular plants}
\format{
A data frame with 24814 observations on the following 2 variables.
\describe{ 
\item{GENUS}{a character vector} 
\item{FAMILY}{a character vector} }
}
\source{
the Plant List Website http://www.theplantlist.org/1.1/browse/-/-/,
with modifications.
}
\description{
A curated genera-family database for vascular plants mainly based on the
Plant List Website http://www.theplantlist.org/1.1/browse/-/-/
}
\examples{

data(genera_dat)

}
\keyword{datasets}
