\name{make_checklist}
\alias{make_checklist}

\title{
Generate a Markdown Checklist of Species based on the output of CTPL
}
\description{
Generated a checklist of all the species based on the output of CTPL, with all the Families appear based on modern classifition systems (eg. APGIII). 
}
\usage{
make_checklist(checklist_dat = NULL, 
                outfile = "output_make_checklist.md", 
                theme = c("complex", "simple", "minimal"))
}

\arguments{
  \item{checklist_dat}{
A dataframe generated by CTPL, or the same form.
}
  \item{outfile}{
The name of the markdown file to generate.
}
  \item{theme}{
Different themes, only selected one once.
}
}
\details{
The output is in Markdown format, which is very convinient to be converted to tex, word, html, pdf ... using Pandoc or the rmarkdown package. 
}
\value{
This function only save the souce code of a checklist of plants to the working directory. It does not return any value.
}
\references{
More details see the vignette.
}
\author{
Jinlong Zhang
}

\seealso{
\code{\link{CTPL}}, \code{\link{count_taxa}}, 
}
\examples{
# Do not run
# For more information, see the vignette
}
