% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantlist-package.R
\docType{package}
\name{plantlist-package}
\alias{plantlist-package}
\alias{plantlist}
\title{Looking Up the Status of Plant Scientific Names based on The Plant List
Database}
\description{
Looking up the status of scientific names based on the Plant List Database
and search the Families for genus.
}
\details{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"plantlist")\Sexpr{tools:::Rd_package_DESCRIPTION("plantlist")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"plantlist")\Sexpr{tools:::Rd_package_indices("plantlist")}
}
\examples{


TPL("Carex")
TPL("Cherry")
splist <- c("Ranunculus japonicus", "Hepatica henryi",
            "Heracleum acuminatum", "Solanum nigrum",
            "Punica sp.", "Machilus", "Today", "####" )
res <- TPL(splist)
taxa.table(res)


}
\author{
Jinlong Zhang

Maintainer: Jinlong Zhang <jinlongzhang01@gmail.com>
}
\keyword{package}
