\name{plantlist-package}
\alias{plantlist-package}
\alias{plantlist}
\docType{package}
\title{
\packageTitle{plantlist}
}
\description{
\packageDescription{plantlist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plantlist}
\packageIndices{plantlist}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{plantlist}

Maintainer: \packageMaintainer{plantlist}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
