\name{taxa.table}
\alias{taxa.table}

\title{
Making a Taxa Table based on the result of function TPL for Phylomatic 
}
\description{
Making taxa table for software Phylomatic, as "Family/Genus/species"
}
\usage{
taxa.table(x, file = NULL, substitute.sp.white.space = "_")
}

\arguments{
  \item{x}{
A data.frame generated by function \code{\link{TPL}} 
}
  \item{file}{
A character string referring the name of the plain text file to be generated. 
}
  \item{substitute.sp.white.space}{
The substitution symbol between the genus and specific epithet. 
}
}
\details{
x must be a dataframe generated by function \code{\link{TPL}}.
}
\value{
Taxa table for software Phylomatic.
}
\references{
Webb, C. O. and M. J. Donoghue (2005). Phylomatic: tree assembly for applied phylogenetics. Molecular Ecology Notes 5(1): 181-183.
}
\author{
Jinlong Zhang \email{ jinlongzhang01@gmail.com }
}

\seealso{
\code{\link{TPL}} 
}
\examples{

sp <- c( "Ranunculus japonicus", "Anemone udensis", 
         "Ranunculus repens", "Ranunculus chinensis", 
         "Solanum nigrum", "Punica sp." ) 
res <- TPL(sp)
taxa.table(res)

}

