\name{TPL}
\alias{TPL}

\title{
Looking up Family and Order at Kew and the plantlist Database
}
\description{
This function enables the user to search the Family (for Angiosperms, Gymnosperms, Ferns and Moss) against Kew Synonym Database, and The plantlist database. 
}
\usage{
TPL(plant.names = NULL)
}

\arguments{
  \item{plant.names}{
 A characteristic vector containing the species or genus of flowering plants. 
}
}
\details{
This function looks up the input character string in extend Kew family list, and tries to find the synonym of the genus. Output is generated by matching the genenus with a local database containing genera and families downloaded from the plantlist Website.
}
\value{
A data frame containing Genus, Kew Genus, Kew Family, The plantlist Family (for flowering plants it is APGIII) and the APGIII orders. 
}
\note{
A small proportion of genus has been divided into several genera, even being placed into different families, this will lead to multiple entries for one enquiry. Pay attention to these issues.
}
\references{
Bremer, Birgitta, K. Bremer, Mark Chase, Mike Fay, James Reveal, Douglas Soltis, Pamella Soltis and Peter Stevens. (2009) "An update of the Angiosperm Phylogeny Group classification for the orders and families of flowering plants: APG III." Botanical Journal of the Linnean Society.

Haston, E., Richardson, J. E., Stevens, P. F., Chase, M. W. and Harris, D. J. (2009). The Linear Angiosperm Phylogeny Group (LAPG) III: a linear sequence of the families in APG III. Botanical Journal of the Linnean Society, 161(2), 128-131.

}
\author{
Jinlong Zhang \email{ jinlongzhang01@gmail.com }
}

\seealso{
\code{\link{taxa.table}}, for generating taxa table for Phylomatic.
}
\examples{

TPL("Carex")
TPL("Apple")
splist <- c( "Ranunculus japonicus", "Solanum nigrum", "Punica sp.", 
             "Machilus", "Today", "####" ) 
res <- TPL(splist)
taxa.table(res)

}
